+++ addons/yat/src/enums/ESceneChangeFailureReason.cs
@@ -1,0 +1,8 @@
+namespace YAT.Enums;
+
+public enum ESceneChangeFailureReason
+{
+	CantOpen,
+	DoesNotExist,
+	CantInstantiate,
+}